+++ .ci/tritonbench/test-cpu.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+set -x
+
+python test/test_cpu/test_cpu.py
