+++ .ci/tritonbench/test-gpu.sh
