+++ test/test_cpu/__init__.py
@@ -1,0 +1,1 @@
+from .test import load_cpu_tests
