+++ test/test_gpu/__init__.py
@@ -1,0 +1,1 @@
+from .test import load_gpu_tests
