+++ test/test_gpu/test.py
@@ -1,0 +1,3 @@
+
+def load_gpu_tests():
+    pass
